@@ -16,8 +16,9 @@
     "bugs": {
         "url": "https://github.com/code-soup/wordpress-plugin-boilerplate/issues"
     },
+    "type": "module",
     "engines": {
-        "node": "^20.0.0"
+        "node": ">=22.0.0"
     },
     "authors": [
         {
@@ -32,59 +33,59 @@
         }
     ],
     "browserslist": [
-        "> 1%",
-        "last 3 versions",
-        "android 4",
-        "safari 10",
-        "not ie <= 8"
+        "> 0.5%",
+        "last 2 versions",
+        "not dead"
     ],
     "scripts": {
         "dev": "webpack serve --hot --mode development --color --config src/config/config.webpack.js",
-        "build": "webpack --mode development --progress --config src/config/config.webpack.js",
-        "build:prod": "webpack --mode production --progress --config src/config/config.webpack.js",
+        "build": "webpack --mode production --progress --config src/config/config.webpack.js",
+        "build:dev": "webpack --mode development --progress --config src/config/config.webpack.js",
+        "analyze": "webpack --mode production --progress --config src/config/config.webpack.js --analyze",
+        "analyze:dev": "webpack --mode development --progress --config src/config/config.webpack.js --analyze",
         "lint": "npm run -s lint:scripts && npm run -s lint:styles",
-        "lint:scripts": "eslint src/scripts src/webpack",
+        "lint:scripts": "eslint src/scripts src/config",
         "lint:styles": "stylelint \"src/styles/**/*.{css,sass,scss}\"",
         "reinit": "rimraf dist && rimraf node_modules && yarn",
         "clean": "rimraf dist"
     },
     "devDependencies": {
-        "@babel/core": "^7.26.0",
-        "@babel/plugin-proposal-class-properties": "^7.18.6",
-        "@babel/preset-env": "^7.26.0",
-        "babel-loader": "^9.2.1",
-        "breakpoint-sass": "^3.0.0",
-        "clean-webpack-plugin": "^4.0.0",
+        "@babel/core": "^7.27.4",
+        "@babel/plugin-transform-class-properties": "^7.27.1",
+        "@babel/preset-env": "^7.27.2",
+        "babel-loader": "^10.0.0",
         "css-loader": "^7.1.2",
-        "cssnano": "^7.0.6",
-        "dotenv": "^16.4.5",
-        "eslint": "^9.15.0",
+        "css-minimizer-webpack-plugin": "^7.0.2",
+        "dotenv": "^16.5.0",
+        "eslint": "^9.28.0",
         "eslint-plugin-import": "^2.31.0",
-        "eslint-webpack-plugin": "^4.2.0",
+        "eslint-webpack-plugin": "^5.0.1",
+        "globals": "^16.2.0",
         "import-glob": "^1.5.0",
         "jquery": "^3.7.1",
         "mini-css-extract-plugin": "^2.9.2",
-        "mini-svg-data-uri": "^1.4.4",
-        "postcss": "^8.4.49",
+        "postcss": "^8.5.4",
         "postcss-loader": "^8.1.1",
-        "postcss-preset-env": "^10.1.0",
+        "postcss-preset-env": "^10.2.1",
         "resolve-url-loader": "^5.0.0",
         "rimraf": "^6.0.1",
-        "sass": "^1.81.0",
-        "sass-loader": "^16.0.3",
+        "sass": "^1.89.1",
+        "sass-loader": "^16.0.5",
         "style-loader": "^4.0.0",
-        "stylelint": "^16.10.0",
+        "stylelint": "^16.20.0",
         "stylelint-config-property-sort-order-smacss": "^10.0.0",
-        "stylelint-config-standard-scss": "^13.1.0",
-        "stylelint-order": "^6.0.4",
+        "stylelint-config-standard-scss": "^15.0.1",
+        "stylelint-order": "^7.0.0",
         "stylelint-webpack-plugin": "^5.0.1",
-        "webpack": "^5.96.1",
-        "webpack-assets-manifest": "^5.2.1",
-        "webpack-cli": "^5.1.4",
-        "webpack-dev-server": "^5.1.0",
+        "svg-spritemap-webpack-plugin": "^4.7.0",
+        "webpack": "^5.99.9",
+        "webpack-bundle-analyzer": "^4.10.2",
+        "webpack-cli": "^6.0.1",
+        "webpack-dev-server": "^5.2.2",
+        "webpack-manifest-plugin": "^5.0.1",
         "webpack-merge": "^6.0.1",
-        "yargs": "~17.7.2",
-        "yarn-audit-fix": "^10.1.0"
+        "yargs": "~18.0.0",
+        "yarn-audit-fix": "^10.1.1"
     },
     "dependencies": {},
     "packageManager": "yarn@1.22.22"
