@@ -1,4 +1,14 @@
 <?php
+/**
+ * Fired when the plugin is uninstalled.
+ *
+ * @package WPPB
+ */
+
+// If uninstall not called from WordPress, then exit.
+if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	exit;
+}
 
 /**
  * Fired when the plugin is uninstalled.
@@ -18,8 +28,4 @@
  *
  * For more information, see the following discussion:
  * https://github.com/tommcfarlin/WordPress-Plugin-Boilerplate/pull/123#issuecomment-28541913
- *
  */
-
-// If uninstall not called from WordPress, then exit.
-defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
